pyproject.toml
[tool.poetry]
name = "amei-exercises"
version = "0.1.0"
description = ""
authors = ["Michael Berg-Mohnicke <michael.berg@zalf.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
zalfmas-common = "^0.1.21"
pyzmq = "^26.2.0"
pandas = "^2.2.2"
openpyxl = "^3.1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
